requirements.txt
starlette
jinja2
sqlalchemy
websockets
sqlalchemy-utils
aiosqlite
itsdangerous
httpx
twitchio
pyyaml
python-multipart
googletrans==3.1.0a0
langid
bs4